requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.5
decorator==5.1.1
Django==4.0.5
django-braces==1.15.0
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
django-stdimage==6.0.1
executing==0.8.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mpi4py==3.1.3
parso==0.8.3
phonenumbers==8.12.56
phonenumberslite==8.12.56
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.29
pure-eval==0.2.2
Pygments==2.12.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.3.0
tzdata==2022.1
wcwidth==0.2.5
